Fixed environment GYM
Put both the agent and environment in the same class inheriting from openaiGYM. We fixed the case where the agent runs out of availabe actions(by making him remove gems if he ends in this situation)
We tried to implement a DQN agent by flattening and hashing the observation at each step. The issue we encountered is that the action size determines the  amount of actions possible but we don't have all actions available at each step so the index is wrong. We need to try to map our valid actions to indexes in a specific list of actions ideally of size 57.