Minor Changes
Fixed test by setting self.epsilon=0 when testing.
Added comments and average turns when testing.
Added a condition where if the game is a failure the agent replays the game setting the reward for all the steps in the failure episode to a negative reward depending on how late in the game the agent is. In  Not sure if this makes the training better needs more experimenting.